Rajoute get_csv_file_dialect
