+++ common/utils/file_import.py
@@ -1,7 +1,19 @@
+import csv
 import hashlib
+import logging
 
+import chardet
 from django.conf import settings
 from django.core.exceptions import ValidationError
+
+logger = logging.getLogger(__name__)
+
+
+def decode_bytes(bytes_string):
+    detection_result = chardet.detect(bytes_string)
+    encoding_detected = detection_result["encoding"]
+    logger.info(f"Encoding autodetected : {encoding_detected}")
+    return (bytes_string.decode(encoding_detected), encoding_detected)
 
 
 def get_file_size(file):
@@ -32,3 +44,10 @@ def get_file_digest(file):
     for row in file:
         file_hash.update(row)
     return file_hash.hexdigest()
+
+
+def get_csv_file_dialect(file):
+    file.seek(0)
+    row_1 = file.readline()
+    (decoded_row, _) = decode_bytes(row_1)
+    return csv.Sniffer().sniff(decoded_row)
